Resolved Drifting Date Issue
We resolve a situation where I have three recurring events with the same title during one week.
One is on Tues, another on Wed, another on Thurs.
There's a bug that occurs, when the script updates.
It turns the Tuesday event into a Wed event, and it turns the Wed event in a Thursday event.
Leaving us with one Wednesday event and two Thursday events.